Cosmetic edits & add comments
