+++ game/shared/igamesystem.h
@@ -11,8 +11,6 @@
 #pragma once
 #endif
 
-#include <string.h>
-
 struct EventGameInit_t;
 struct EventGameShutdown_t;
 struct EventGamePostInit_t;
@@ -90,7 +88,7 @@ public:
 
 	virtual void ClientPreEntityThink(const EventClientPreEntityThink_t* const msg) = 0;
 
-	virtual void nullsub_1271(const void* const msg) = 0;
+	virtual void unk_1271(const void* const msg) = 0;
 
 	// Gets called each frame
 	virtual void Update(const EventClientUpdate_t* const msg) = 0;
@@ -104,13 +102,13 @@ public:
 	virtual void FrameUpdatePostEntityThink(const EventServerPostEntityThink_t* const msg) = 0;
 	virtual void PreClientUpdate(const EventServerPreClientUpdate_t* const msg) = 0;
 
-	virtual void nullsub_1277(const void* const msg) = 0;
-	virtual void nullsub_1278(const void* const msg) = 0;
+	virtual void unk_1277(const void* const msg) = 0;
+	virtual void unk_1278(const void* const msg) = 0;
 
 	virtual void OnServerGamePostSimulate(const EventServerGamePostSimulate_t* const msg) = 0;
 	virtual void OnClientGamePostSimulate(const EventClientGamePostSimulate_t* const msg) = 0;
 
-	virtual void nullsub_1281(const void* const msg) = 0;
+	virtual void unk_1281(const void* const msg) = 0;
 
 	virtual void FrameBoundary(const EventGameFrameBoundary_t* const msg) = 0;
 	virtual void OutOfGameFrameBoundary(const EventOutOfGameFrameBoundary_t* const msg) = 0;
@@ -118,18 +116,17 @@ public:
 	virtual void SaveGame(const EventSaveGame_t* const msg) = 0;
 	virtual void RestoreGame(const EventRestoreGame_t* const msg) = 0;
 
-	virtual void nullsub_1285(const void* const msg) = 0;
-	virtual void nullsub_1286(const void* const msg) = 0;
-	virtual void nullsub_1287(const void* const msg) = 0;
-	virtual void nullsub_1288(const void* const msg) = 0;
-	virtual void nullsub_1289(const void* const msg) = 0;
+	virtual void unk_1285(const void* const msg) = 0;
+	virtual void unk_1286(const void* const msg) = 0;
+	virtual void unk_1287(const void* const msg) = 0;
+	virtual void unk_1288(const void* const msg) = 0;
+	virtual void unk_1289(const void* const msg) = 0;
 
 	virtual const char* GetName() = 0;
 	virtual void SetGameSystemGlobalPtrs(void* pValue) = 0;
 	virtual void SetName(const char* pName) = 0;
 	virtual bool DoesGameSystemReallocate() = 0;
-	virtual ~IGameSystem() = 0;
-	virtual void YouForgot_DECLARE_GAME_SYSTEM_InYourClassDefinition() = 0;
+	virtual ~IGameSystem() {}
 };
 
 // Quick and dirty server system for users who don't care about precise ordering
@@ -178,7 +175,7 @@ public:
 
 	virtual void ClientPreEntityThink(const EventClientPreEntityThink_t* const msg) {}
 
-	virtual void nullsub_1271(const void* const msg) {}
+	virtual void unk_1271(const void* const msg) {}
 
 	// Gets called each frame
 	virtual void Update(const EventClientUpdate_t* const msg) {}
@@ -192,13 +189,13 @@ public:
 	virtual void FrameUpdatePostEntityThink(const EventServerPostEntityThink_t* const msg) {}
 	virtual void PreClientUpdate(const EventServerPreClientUpdate_t* const msg) {}
 
-	virtual void nullsub_1277(const void* const msg) {}
-	virtual void nullsub_1278(const void* const msg) {}
+	virtual void unk_1277(const void* const msg) {}
+	virtual void unk_1278(const void* const msg) {}
 
 	virtual void OnServerGamePostSimulate(const EventServerGamePostSimulate_t* const msg) {}
 	virtual void OnClientGamePostSimulate(const EventClientGamePostSimulate_t* const msg) {}
 
-	virtual void nullsub_1281(const void* const msg) {}
+	virtual void unk_1281(const void* const msg) {}
 
 	virtual void FrameBoundary(const EventGameFrameBoundary_t* const msg) {}
 	virtual void OutOfGameFrameBoundary(const EventOutOfGameFrameBoundary_t* const msg) {}
@@ -206,11 +203,11 @@ public:
 	virtual void SaveGame(const EventSaveGame_t* const msg) {}
 	virtual void RestoreGame(const EventRestoreGame_t* const msg) {}
 
-	virtual void nullsub_1285(const void* const msg) {}
-	virtual void nullsub_1286(const void* const msg) {}
-	virtual void nullsub_1287(const void* const msg) {}
-	virtual void nullsub_1288(const void* const msg) {}
-	virtual void nullsub_1289(const void* const msg) {}
+	virtual void unk_1285(const void* const msg) {}
+	virtual void unk_1286(const void* const msg) {}
+	virtual void unk_1287(const void* const msg) {}
+	virtual void unk_1288(const void* const msg) {}
+	virtual void unk_1289(const void* const msg) {}
 
 	virtual const char* GetName() { return m_pName; }
 	virtual void SetGameSystemGlobalPtrs(void* pValue) {}
@@ -227,123 +224,5 @@ class CAutoGameSystem : public CBaseGameSystem
 protected:
 	virtual ~CAutoGameSystem() {};
 };
-
-class IGameSystemFactory
-{
-public:
-	virtual bool Init() = 0;
-	virtual void PostInit() = 0;
-	virtual void Shutdown() = 0;
-	virtual IGameSystem* CreateGameSystem() = 0;
-	virtual void DestroyGameSystem(IGameSystem* pGameSystem) = 0;
-	virtual bool ShouldAutoAdd() = 0;
-	virtual int GetPriority() = 0;
-	virtual void SetGlobalPtr(void* pValue) = 0;
-	virtual bool IsReallocating() = 0;
-	virtual IGameSystem* GetStaticGameSystem() = 0;
-};
-
-class CBaseGameSystemFactory : public IGameSystemFactory
-{
-protected:
-	CBaseGameSystemFactory(const char* pName)
-	{
-		m_pName = pName;
-		m_pNext = *CBaseGameSystemFactory::sm_pFirst;
-		*CBaseGameSystemFactory::sm_pFirst = this;
-	}
-private:
-	CBaseGameSystemFactory* m_pNext;
-	const char* m_pName;
-public:
-	static IGameSystem* GetGlobalPtrByName(const char* pName)
-	{
-		CBaseGameSystemFactory* pFactoryList = *CBaseGameSystemFactory::sm_pFirst;
-		while (pFactoryList)
-		{
-			if (strcmp(pFactoryList->m_pName, pName) == 0)
-			{
-				return pFactoryList->GetStaticGameSystem();
-			}
-			pFactoryList = pFactoryList->m_pNext;
-		}
-		return nullptr;
-	}
-	static CBaseGameSystemFactory** sm_pFirst;
-};
-template <class T, class U = T>
-class CGameSystemStaticFactory : public CBaseGameSystemFactory
-{
-public:
-	CGameSystemStaticFactory(const char* pName, T* pActualGlobal, U** ppGlobalPointer = nullptr) : CBaseGameSystemFactory(pName)
-	{
-		m_pActualGlobal = pActualGlobal;
-		m_ppGlobalPointer = ppGlobalPointer;
-		pActualGlobal->SetName(pName);
-	}
-	bool Init() override { return m_pActualGlobal->Init(); }
-	void PostInit() override { m_pActualGlobal->PostInit(); }
-	void Shutdown() override { m_pActualGlobal->Shutdown(); }
-	IGameSystem* CreateGameSystem() override
-	{
-		m_pActualGlobal->SetGameSystemGlobalPtrs(m_pActualGlobal);
-		return m_pActualGlobal;
-	}
-	void DestroyGameSystem(IGameSystem* pGameSystem) override
-	{
-		m_pActualGlobal->SetGameSystemGlobalPtrs(nullptr);
-	}
-	bool ShouldAutoAdd() override { return true; }
-	int GetPriority() override { return 0; }
-	void SetGlobalPtr(void* pValue) override
-	{
-		if (m_ppGlobalPointer)
-		{
-			*m_ppGlobalPointer = reinterpret_cast<T*>(pValue);
-		}
-	}
-	bool IsReallocating() override { return false; }
-	IGameSystem* GetStaticGameSystem() override { return m_pActualGlobal; }
-private:
-	T* m_pActualGlobal;
-	U** m_ppGlobalPointer;
-};
-template <class T, class U = T>
-class CGameSystemReallocatingFactory : public CBaseGameSystemFactory
-{
-public:
-	CGameSystemReallocatingFactory(const char* pName, U** ppGlobalPointer = nullptr) : CBaseGameSystemFactory(pName)
-	{
-		m_ppGlobalPointer = ppGlobalPointer;
-	}
-	bool Init() override { return true; }
-	void PostInit() override { }
-	void Shutdown() override { }
-	IGameSystem* CreateGameSystem() override
-	{
-		T* pObject = new T();
-		pObject->SetGameSystemGlobalPtrs(pObject);
-		return pObject;
-	}
-	void DestroyGameSystem(IGameSystem* pGameSystem) override
-	{
-		pGameSystem->SetGameSystemGlobalPtrs(nullptr);
-		delete pGameSystem;
-	}
-	bool ShouldAutoAdd() override { return true; }
-	int GetPriority() override { return 0; }
-	void SetGlobalPtr(void* pValue) override
-	{
-		if (m_ppGlobalPointer)
-		{
-			*m_ppGlobalPointer = reinterpret_cast<T*>(pValue);
-		}
-	}
-	bool IsReallocating() override { return true; }
-	IGameSystem* GetStaticGameSystem() override { return nullptr; }
-private:
-	U** m_ppGlobalPointer;
-};
-
 
 #endif // IGAMESYSTEM_H
